Deleted gdc.grin2.list and gdc.grin2.run. Renamed grin2.gdc.ts to grin2.ts
